Move helper class ownership
